third_party/WebKit/LayoutTests/editing/pasteboard/paste_before_placeholder.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_not_equals(window.internals, undefined,
        'This test requires window.internals to access clipboard');

    assert_selection(
        '<div contenteditable>^one|<br>two</div>',
        selection => {
            selection.document.execCommand('cut');
            selection.document.execCommand('paste');
        },
        '<div contenteditable>one|<br>two</div>',
        'Paste before placeholder');
});
</script>
